refactor: create chat container for chat component
